:art: moved and renamed BruteForceClaimsFetcher, various renamings and comments
