+++ service/security/shiro/src/main/java/org/eclipse/kapua/service/authorization/shiro/AuthorizationModule.java
@@ -94,9 +94,8 @@ import org.eclipse.kapua.service.authorization.role.shiro.RolePermissionFactoryI
 import org.eclipse.kapua.service.authorization.role.shiro.RolePermissionImplJpaRepository;
 import org.eclipse.kapua.service.authorization.role.shiro.RolePermissionServiceImpl;
 import org.eclipse.kapua.service.authorization.role.shiro.RoleServiceImpl;
-import org.eclipse.kapua.service.authorization.shiro.claims.BruteForceClaimsFetcher;
 import org.eclipse.kapua.service.authorization.shiro.claims.ClaimsFetcher;
-import org.eclipse.kapua.service.authorization.shiro.claims.SimpleClaimsFetcher;
+import org.eclipse.kapua.service.authorization.shiro.claims.ClaimsFetcherImpl;
 import org.eclipse.kapua.service.authorization.shiro.setting.KapuaAuthorizationSetting;
 import org.eclipse.kapua.storage.TxManager;
 
@@ -202,26 +201,13 @@ public class AuthorizationModule extends AbstractKapuaModule {
 
     @Provides
     @Singleton
-    @Named("bruteForceClaimsFetcher")
-    ClaimsFetcher bruteForceClaimsFetcher(
-            AuthorizationService authorizationService,
-            PermissionFactory permissionFactory,
-            Set<Domain> knownDomains) {
-        return new BruteForceClaimsFetcher(
-                authorizationService,
-                permissionFactory,
-                knownDomains);
-    }
-
-    @Provides
-    @Singleton
     @Named("defaultClaimsFetcher")
     ClaimsFetcher defaultClaimsFetcher(
             AuthorizationService authorizationService,
             AuthenticationService authenticationService,
             PermissionFactory permissionFactory,
             Set<Domain> knownDomains) {
-        return new SimpleClaimsFetcher(
+        return new ClaimsFetcherImpl(
                 authorizationService,
                 authenticationService,
                 permissionFactory,
